requirements.txt
asgiref==3.5.0
Django==4.0.3
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap3-datetimepicker==2.2.3
django-computed-property==0.3.0
django-datetime-widget2==0.9.5
django-widget-tweaks==1.4.12
Pillow==9.0.1
PyMySQL==1.0.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
